+++ pyvela/examples/run_example_emcee.py
@@ -1,58 +1,60 @@
 #!/usr/bin/env python
 
 # %%
-import sys
 from timeit import timeit
 
 import corner
 import emcee
 import numpy as np
 from matplotlib import pyplot as plt
-from pint.logging import setup as setup_log
-from pint.models import get_model_and_toas
+from argparse import ArgumentParser
 
-from pyvela import Vela as vl
-from pyvela import read_model_and_toas
+from pyvela import SPNTA, Vela as vl
+from pyvela.priors import parse_custom_prior_file
 
-# %%
-setup_log(level="WARNING")
 
-# %%
-parfile, timfile = sys.argv[1], sys.argv[2]
-m, t = get_model_and_toas(parfile, timfile)
+parser = ArgumentParser()
+parser.add_argument("par_file")
+parser.add_argument("tim_file")
+parser.add_argument("-P", "--prior_file")
+args = parser.parse_args()
 
-# %%
-mv, tv = read_model_and_toas(parfile, timfile, cheat_prior_scale=10)
-lnpost = vl.get_lnpost_func(mv, tv, True)
-prior_transform = vl.get_prior_transform_func(mv)
 
 # %%
-param_names = vl.get_free_param_names(mv.param_handler)
-scale_factors = vl.get_scale_factors(mv.param_handler)
-shifts = [(float(m.F0.value) if pname == "F0" else 0) for pname in param_names]
-
-# %%
-maxlike_params_v = np.array(
-    [
-        vl.read_param_values_to_vector(
-            mv.param_handler, mv.param_handler._default_params_tuple
-        )
-    ]
+prior_dict = (
+    parse_custom_prior_file(args.prior_file) if args.prior_file is not None else {}
 )
 
 # %%
-print(lnpost(maxlike_params_v))
-print(timeit("lnpost(maxlike_params_v)", globals=globals(), number=1000))
+spnta = SPNTA(
+    args.par_file, args.tim_file, cheat_prior_scale=10, custom_priors=prior_dict
+)
 
 # %%
-ndim = len(param_names)
-nwalkers = ndim * 5
-p0 = np.array([prior_transform(cube) for cube in np.random.rand(nwalkers, ndim)])
+shifts = [
+    (spnta.model.param_handler._default_params_tuple.F_.x if pname == "F0" else 0)
+    for pname in spnta.param_names
+]
+
+# %%
+maxlike_params_v = np.array([spnta.maxlike_params])
+
+# %%
+print(spnta.lnpost_vectorized(maxlike_params_v))
+print(
+    timeit("spnta.lnpost_vectorized(maxlike_params_v)", globals=globals(), number=1000)
+)
+
+# %%
+nwalkers = spnta.ndim * 5
+p0 = np.array(
+    [spnta.prior_transform(cube) for cube in np.random.rand(nwalkers, spnta.ndim)]
+)
 
 sampler = emcee.EnsembleSampler(
     nwalkers,
-    ndim,
-    lnpost,
+    spnta.ndim,
+    spnta.lnpost_vectorized,
     moves=[emcee.moves.StretchMove(), emcee.moves.DESnookerMove()],
     vectorize=True,
 )
@@ -60,7 +62,7 @@ sampler.run_mcmc(p0, 6000, progress=True, progress_kwargs={"mininterval": 1})
 
 # %%
 samples_v_0 = sampler.get_chain(flat=True, discard=2500, thin=40)
-samples_v = samples_v_0 / scale_factors
+samples_v = spnta.rescale_samples(samples_v_0)
 
 # %%
 params_no_plot = [
@@ -73,14 +75,14 @@ params_no_plot = [
 ]
 param_plot_mask = [
     idx
-    for idx, par in enumerate(param_names)
+    for idx, par in enumerate(spnta.param_names)
     if all(not par.startswith(pnp) for pnp in params_no_plot)
 ]
 
 # %%
-means = (np.mean(samples_v_0, axis=0) + shifts) / scale_factors
+means = (np.mean(samples_v_0, axis=0) + shifts) / spnta.scale_factors
 stds = np.std(samples_v, axis=0)
-for idx, (pname, mean, std) in enumerate(zip(param_names, means, stds)):
+for idx, (pname, mean, std) in enumerate(zip(spnta.param_names, means, stds)):
     if idx in param_plot_mask:
         print(f"{pname}\t\t{mean}\t\t{std}")
 
@@ -88,7 +90,7 @@ for idx, (pname, mean, std) in enumerate(zip(param_names, means, stds)):
 # param_labels = [f"\n\n{pname}\n({m[pname].units})\n" for pname in param_names]
 param_labels = [
     f"\n\n{label}\n\n"
-    for idx, label in enumerate(vl.get_free_param_labels(mv))
+    for idx, label in enumerate(spnta.param_labels)
     if idx in param_plot_mask
 ]
 samples_for_plot = samples_v[:, param_plot_mask]
@@ -97,30 +99,36 @@ fig = corner.corner(
     labels=param_labels,
     label_kwargs={"fontsize": 11},
     range=[0.999] * len(param_labels),
-    truths=(maxlike_params_v[0] / scale_factors)[param_plot_mask],
+    truths=(maxlike_params_v[0] / spnta.scale_factors)[param_plot_mask],
     plot_datapoints=False,
     hist_kwargs={"density": True},
 )
 
-plt.suptitle(m.PSR.value)
+plt.suptitle(spnta.model.pulsar_name)
 plt.tight_layout()
 plt.show()
 
 # %%
-params_median = vl.read_params(mv, np.median(samples_v_0, axis=0))
+params_median = vl.read_params(spnta.model, np.median(samples_v_0, axis=0))
 rv = (
-    list(map(vl.value, vl.form_residuals(mv, tv, params_median)))
-    if not t.is_wideband()
-    else [vl.value(wr[0]) for wr in vl.form_residuals(mv, tv, params_median)]
+    list(map(vl.value, vl.form_residuals(spnta.model, spnta.toas, params_median)))
+    if not spnta.is_wideband()
+    else [
+        vl.value(wr[0])
+        for wr in vl.form_residuals(spnta.model, spnta.toas, params_median)
+    ]
 )
 
-ctoas = [vl.correct_toa(mv, tvi, params_median) for tvi in tv]
+ctoas = [vl.correct_toa(spnta.model, tvi, params_median) for tvi in spnta.toas]
 errs = np.sqrt(
-    [vl.value(vl.scaled_toa_error_sqr(tvi, ctoa)) for (tvi, ctoa) in zip(tv, ctoas)]
+    [
+        vl.value(vl.scaled_toa_error_sqr(tvi, ctoa))
+        for (tvi, ctoa) in zip(spnta.toas, ctoas)
+    ]
 )
 
 plt.errorbar(
-    t.get_mjds(),
+    spnta.get_mjds(),
     rv,
     errs,
     ls="",
@@ -130,6 +138,6 @@ plt.errorbar(
 plt.axhline(0, color="grey", ls="dotted")
 plt.xlabel("MJD")
 plt.ylabel("Residuals (s)")
-plt.suptitle(m.PSR.value)
+plt.suptitle(spnta.model.pulsar_name)
 plt.tight_layout()
 plt.show()