+++ pyvela/tests/test_data_files.py
@@ -4,7 +4,6 @@ import pytest
 from juliacall import Main as jl
 from pint.models import get_model_and_toas
 
-from pyvela import read_model_and_toas
 from pyvela.model import fix_red_noise_components
 from pyvela.spnta import SPNTA
 from pyvela.vela import vl
@@ -51,7 +50,7 @@ datasets = [
 ]
 
 
-@pytest.fixture(params=datasets, scope="module")
+@pytest.fixture(params=datasets)
 def model_and_toas(request):
     dataset = request.param