+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - `pyvela.SPNTA` class
 ## Changed
 - Reorganized `pyvela` code
+- Rerun failures in `pyvela` CI tests.
+- Example scripts to use `pyvela.SPNTA`
 ## Fixed
 ## Removed